Pasar el contenido de cdp_main.py a main.py
